Rename object-storage-integrator to azure-storage-integrator
