+++ .github/workflows/sync_issue_to_jira.yaml
@@ -1,5 +1,6 @@
-# Copyright 2023 Canonical Ltd.
+# Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
+
 name: Sync issue to Jira
 
 on:
