+++ charmcraft.yaml
@@ -17,6 +17,3 @@ bases:
   - name: ubuntu
     channel: "22.04"
     architectures: [amd64]
-  # - name: ubuntu
-  #   channel: "22.04"
-  #   architectures: [arm64]
