+++ src/events/provider.py
@@ -5,7 +5,7 @@
 """Azure Storage Provider related event handlers."""
 
 
-from charms.data_platform_libs.v0.object_storage import (
+from charms.data_platform_libs.v0.azure_storage import (
     AzureStorageProviderData,
     AzureStorageProviderEventHandlers,
     StorageConnectionInfoRequestedEvent,
@@ -15,7 +15,7 @@ from ops import CharmBase
 from constants import AZURE_RELATION_NAME
 from core.context import Context
 from events.base import BaseEventHandler
-from managers.object_storage import ObjectStorageManager
+from managers.azure_storage import AzureStorageManager
 from utils.logging import WithLogging
 
 
@@ -32,7 +32,7 @@ class AzureStorageProviderEvents(BaseEventHandler, WithLogging):
         self.azure_provider = AzureStorageProviderEventHandlers(
             self.charm, self.azure_provider_data
         )
-        self.object_storage_manager = ObjectStorageManager(self.azure_provider_data)
+        self.azure_storage_manager = AzureStorageManager(self.azure_provider_data)
 
         self.framework.observe(
             self.azure_provider.on.storage_connection_info_requested,
@@ -49,4 +49,4 @@ class AzureStorageProviderEvents(BaseEventHandler, WithLogging):
         container_name = self.charm.config.get("container")
         assert container_name is not None
 
-        self.object_storage_manager.update(self.context.azure_storage)
+        self.azure_storage_manager.update(self.context.azure_storage)
