+++ tests/integration/test-charm-azure/charmcraft.yaml
@@ -6,9 +6,6 @@ bases:
   - name: ubuntu
     channel: "22.04"
     architectures: [amd64]
-  # - name: ubuntu
-  #   channel: "22.04"
-  #   architectures: [arm64]
 
 parts:
   charm:
